[upd] google-anyscale-cloudapis module - remove sts api
The STS api is not required for Anyscale to function. When you create
the roles on the project, you can use the `roles/iam.serviceAccountTokenCreator`
and that will allow the Anyscale service account to create tokens for
itself. When doing this from the Google Cloud Console, it automatically
enables the STS API, but it is not required.

On branch release-0.2.4
Changes to be committed:
	modified:   modules/google-anyscale-cloudapis/README.md
	modified:   modules/google-anyscale-cloudapis/variables.tf