[upd] google-anyscale-filestore submodule - change default tier/size
This commit updates the default tier and size for the filestore. It matches
the defaults used in `anyscale cloud setup`. The default tier is now `STANDARD`
and the default size is now 1024GB.

This also required changing the default location from region to availability zone.
Because not all regions have a default AZ-A, the default location is now set to
the region and AZ-B.

On branch brent/filestore-defaults
Changes to be committed:
	modified:   CHANGELOG.md
	modified:   README.md
	modified:   examples/anyscale-v2-commonname/main.tf
	modified:   examples/anyscale-v2-existingproject/main.tf
	new file:   examples/anyscale-v2-kitchensink/README.md
	modified:   main.tf
	modified:   modules/google-anyscale-filestore/README.md
	modified:   modules/google-anyscale-filestore/main.tf
	modified:   modules/google-anyscale-filestore/variables.tf
	modified:   variables.tf