feat: Add logging sink module
Add a module for managing the _Default logging sink in Google Cloud Logging.
The default configuration for Google Cloud Logging is to enable syslog for all
compute instances. This can cause a high volume of logs leading to unexpected
costs. This optional module allows you to disable that behavior.

Additional updates include changes to the pre-commit configuration to the
latest version, updates to tflint to the latest version, cleanups for the
trivy linter, and updates to readme's and changelog.

On branch brent/logging-sink
Changes to be committed:
	modified:   .pre-commit-config.yaml
	modified:   .tflint.hcl
	modified:   CHANGELOG.md
	modified:   README.md
	modified:   examples/anyscale-v2-commonname/main.tf
	modified:   main.tf
	modified:   modules/google-anyscale-cloudstorage/main.tf
	new file:   modules/google-anyscale-loggingsink/README.md
	new file:   modules/google-anyscale-loggingsink/data.tf
	new file:   modules/google-anyscale-loggingsink/main.tf
	new file:   modules/google-anyscale-loggingsink/outputs.tf
	new file:   modules/google-anyscale-loggingsink/variables.tf
	new file:   modules/google-anyscale-loggingsink/versions.tf
	modified:   variables.tf