fix: VPC Firewall when using ingress_from_cidr_map
One of the options available from the module is to use the ingress_from_cidr_map to allow traffic from specific CIDR ranges.
This was not working when passing in a single port and cidr range.

Additional update to the KitchenSink example to test this use.

On branch brent/vpc-fixfirewall
Changes to be committed:
	modified:   README.md
	modified:   examples/README.md
	modified:   examples/main.tf
	modified:   main.tf
	modified:   variables.tf