fix: IAM SA Desc when passing in a CloudID
When passing in an Cloud ID, the SA description was trying to add var.google_region however
this variable was always null. This caused the IAM SA to fail to create.

Additional updates to pre-commit-config.yaml to update to the latest hook versions.

On branch brent/fix-iam-sa-desc
Changes to be committed:
	modified:   .pre-commit-config.yaml
	modified:   modules/google-anyscale-iam/README.md
	new file:   modules/google-anyscale-iam/data.tf
	modified:   modules/google-anyscale-iam/gke-sa.tf