fix: google-anyscale-iam: IAM SA Desc when passing in a CloudID (#42)
* fix: IAM SA Desc when passing in a CloudID

When passing in an Cloud ID, the SA description was trying to add var.google_region however
this variable was always null. This caused the IAM SA to fail to create.

Additional updates to pre-commit-config.yaml to update to the latest hook versions.

On branch brent/fix-iam-sa-desc
Changes to be committed:
	modified:   .pre-commit-config.yaml
	modified:   modules/google-anyscale-iam/README.md
	new file:   modules/google-anyscale-iam/data.tf
	modified:   modules/google-anyscale-iam/gke-sa.tf

* upd: CHANGELOG to reflect recent changes.

This hasn't been updated in a while. I pulled changes
from release descriptions and updated.

Changes to be committed:
	modified:   CHANGELOG.md

* fix: Additional description updates

Additional updates to use the data source for region when a cloud id is passed in.

Changes to be committed:
	modified:   ../../modules/google-anyscale-iam/gke-sa.tf
	modified:   ../../modules/google-anyscale-iam/main.tf